# src/modules/course/course.route.ts
// course.route.ts
import { NextFunction, Request, Response, Router } from 'express';
import { courseController } from './course.controller';
import { validateData } from '../../middlewares/validateData';
import { createCourseSchema, updateCourseSchema } from './course.validation';
import { upload } from '../../utils/uploadFile';

const router = Router();

router.post(
  '/',
  /**
   * #swagger.summary = 'Create a new course'
   * #swagger.description = 'Create a new course by passing course information in the request body.'
   * #swagger.parameters['file'] = {
      in: 'formData',
      type: 'file',
      required: true,
      description: 'Course thumbnail image'
    }
    #swagger.parameters['course_info'] = {
      in: 'formData',
      type: 'string',
      required: true,
      description: 'Course info as JSON string',
    }
   */
  upload.single('file'),
  (req: Request, res: Response, next: NextFunction) => {
    if (!req.file) {
      return res.status(400).json({ message: 'File is required' });
    }
    req.body = { ...JSON.parse(req.body?.course_info), thumbnail: `/uploads/${req.file.filename}` };
    next();
  },
  validateData(createCourseSchema),
  courseController.createCourse
);

router.get('/', courseController.getAllCourses);
router.get('/:id', courseController.getCourseById);

router.put('/:id', validateData(updateCourseSchema), courseController.updateCourse);

router.delete('/:id', courseController.deleteCourse);

export const courseRoutes = router;