# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My PERN API",
    "version": "1.0.0",
    "description": "PERN Stack API Documentation"
  },
  "host": "http://localhost:3000",
  "basePath": "",
...
  "schemes": [],
  "securityDefinitions": {},
  "paths": {
...
    "/api/v1/courses/": {
      "post": {
        "tags": [
          "Courses"
        ],
        "summary": "Create a new course",
        "description": "Create a new course by passing course information in the request body.",
        "parameters": [
          {
            "name": "file",
            "in": "formData",
            "type": "file",
            "required": true,
            "description": "Course thumbnail image"
          },
          {
            "name": "course_info",
            "in": "formData",
            "type": "string",
            "required": true,
            "description": "Course info as JSON string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "get": {
        "tags": [
          "Courses"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/courses/{id}": {
      "get": {
        "tags": [
          "Courses"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Courses"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Courses"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  }
}